Add files via upload
Editing section on correcting scanning errors following R exams instructions. No longer need to go through the merged PDF file with all student evaluations. The instructions for merging the HTML files into a single PDF are still in the manual because it remains a convenient file for a quick inspection. These instructions were moved to the end of the manual now, in the "Troubleshooting and other bits and pieces" section.